requirements.txt
djangorestframework
django-cors-headers
Pillow